improved performance by removing components that re-render, implemented component Lifecycle Methods and restructured folder structure
