+++ src/components/Burger/Burger.js
@@ -18,7 +18,7 @@ const burger = (props) => {
       .reduce((arr, el) => {
         return arr.concat(el)
       }, []);
-      console.log(transformedIngredients);
+  //    console.log(transformedIngredients);
   if (transformedIngredients.length === 0) {
     transformedIngredients = <p>Please start adding Ingredients</p>
   }
