+++ src/components/Burger/OrderSummary/OrderSummary.js
@@ -1,31 +1,38 @@
-import React from 'react';
+import React, { Component } from 'react';
 
-import Aux from '../../../hoc/Aux';
+import Aux from '../../../hoc/Aux/Aux';
 import Button from '../../UI/Button/Button';
 
-const orderSummary = (props) => {
-  const ingredientSummary = Object.keys(props.ingredients)
-      .map(igKey => {
-          return (
-                    <li key={igKey}>
-                        <span style={{textTransform: 'capitalize'}}>{igKey}</span>: {props.ingredients[igKey]}
-                    </li>
-                  );
-      });
-  return (
-    <Aux>
-        <h3>Your Order</h3>
-        <p>A delicious burger with the following ingredients:</p>
-        <ul>
-            {ingredientSummary}
-        </ul>
-        <p><strong>Total Price: {props.price.toFixed(2)}</strong></p>
-        <p>Continue to Checkout?</p>
-        <Button btnType="Danger" clicked={props.purchaseCancelled}>CANCEL</Button>
-        <Button btnType="Success" clicked={props.purchaseContinued}>CONTINUE</Button>
+class OrderSummary  extends Component {
+  //This could be a functional component, desent have to be a class
 
-    </Aux>
-  )
+  componentWillUpdate() {
+    console.log('[OrderSummary] WillUpdate');
+  }
+    render () {
+      const ingredientSummary = Object.keys(this.props.ingredients)
+          .map(igKey => {
+              return (
+                        <li key={igKey}>
+                            <span style={{textTransform: 'capitalize'}}>{igKey}</span>: {this.props.ingredients[igKey]}
+                        </li>
+                      );
+          });
+      return (
+        <Aux>
+            <h3>Your Order</h3>
+            <p>A delicious burger with the following ingredients:</p>
+            <ul>
+                {ingredientSummary}
+            </ul>
+            <p><strong>Total Price: {this.props.price.toFixed(2)}</strong></p>
+            <p>Continue to Checkout?</p>
+            <Button btnType="Danger" clicked={this.props.purchaseCancelled}>CANCEL</Button>
+            <Button btnType="Success" clicked={this.props.purchaseContinued}>CONTINUE</Button>
+
+        </Aux>
+      )
+    }
 };
 
-export default orderSummary;
+export default OrderSummary;
