+++ src/components/Navigation/SideDrawer/DrawerToggle/DrawerToggle.js
@@ -6,10 +6,7 @@ const drawerToggle = (props) => (
     <div className={classes.DrawerToggle} onClick={props.clicked}>
         <div></div>
         <div></div>
-        <div>Burger</div>
         <div></div>
-        <div></div>
-
     </div>
 );
