+++ src/components/UI/Modal/Modal.js
@@ -1,21 +1,34 @@
-import React from 'react';
+import React, { Component } from 'react';
 
 import classes from './Modal.css';
-import Aux from '../../../hoc/Aux';
+import Aux from '../../../hoc/Aux/Aux';
 import Backdrop from '../Backdrop/Backdrop';
 
-const modal = (props) => (
-  <Aux>
-    <Backdrop show={props.show} clicked={props.modalClosed} />
-      <div
-          className={classes.Modal}
-          style={{
-            transform: props.show ? 'translateY(0)' : 'translateY(-100vh)',
-            opacity: props.show ? '1': '0'
-          }}>
-          {props.children}
-      </div>
-  </Aux>
-);
+class Modal extends Component {
 
-export default modal;
+  shouldComponentUpdate(nextProps, nextState) {
+      return nextProps.show !== this.props.show;
+  }
+
+  componentWillUpdate () {
+    console.log('[Modal] WillUpdate');
+  }
+
+  render () {
+    return (
+      <Aux>
+        <Backdrop show={this.props.show} clicked={this.props.modalClosed} />
+          <div
+              className={classes.Modal}
+              style={{
+                transform: this.props.show ? 'translateY(0)' : 'translateY(-100vh)',
+                opacity: this.props.show ? '1': '0'
+              }}>
+              {this.props.children}
+          </div>
+      </Aux>
+    );
+  }
+}
+
+export default Modal;
